third_party/WebKit/LayoutTests/external/wpt/css/css-variables/variable-substitution-shadow-properties-expected.txt

This is a testharness.js-based test.
FAIL box-shadow assert_equals: Expected Value should match actual value expected "1px 1px 1px 1px rgb(0,128,0)" but got "rgb(0, 128, 0) 1px 1px 1px 1px"
FAIL text-shadow assert_equals: Expected Value should match actual value expected "1px 1px 1px rgb(0,128,0)" but got "rgb(0, 128, 0) 1px 1px 1px"
Harness: the test ran to completion.
